Updated Sign-up button and link stylings, Moved sign-in form into a component
